refactor(ai-worker): mover executor raiz do GeraLanding para pacote comum
