+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/GeraLandingExecutionScheduler.java
@@ -1,5 +1,6 @@
 package com.marketinghub.worker.geralanding;
 
+import com.marketinghub.worker.geralanding.comum.GeraLandingExecutionService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
