+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/copy/GeraLandingExecutionService.java
@@ -1,21 +1,23 @@
 package com.marketinghub.worker.geralanding.copy;
 
+import com.marketinghub.worker.geralanding.comum.GeraLandingStageExecutionProcessor;
+import com.marketinghub.worker.geralanding.comum.GeraLandingStageExecutionRef;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
 /** Responsabilidade: encapsular execução da etapa copy preservando isolamento de pacote. */
 @Service("geraLandingCopyExecutionStageService")
 public class GeraLandingExecutionService {
-    private final com.marketinghub.worker.geralanding.comum.GeraLandingStageExecutionProcessor executionProcessor;
+    private final GeraLandingStageExecutionProcessor executionProcessor;
 
-    public GeraLandingExecutionService(com.marketinghub.worker.geralanding.comum.GeraLandingStageExecutionProcessor executionProcessor) {
+    public GeraLandingExecutionService(GeraLandingStageExecutionProcessor executionProcessor) {
         this.executionProcessor = executionProcessor;
     }
 
     /** Processa as execuções da etapa copy convertendo para DTO base do pipeline. */
     public void processExecutions(List<GeraLandingStageExecutionDto> jobs) {
         executionProcessor.processExecutions(jobs.stream()
-                .map(item -> new com.marketinghub.worker.geralanding.comum.GeraLandingStageExecutionRef(item.experimentId(), item.idJob(), item.stageCode()))
+                .map(item -> new GeraLandingStageExecutionRef(item.experimentId(), item.idJob(), item.stageCode()))
                 .toList());
     }
 }
